Log part and appInfo part.
