+++ LLDebugTool.podspec
@@ -19,8 +19,21 @@ Pod::Spec.new do |s|
 #    ss.public_header_files      = "LLDebugTool/Core/Component/Network/LLNetwork.h"
 #    ss.source_files             = "LLDebugTool/DebugTool/*.{h,m}", "LLDebugTool/Core/Component/Network/**/*.{h,m}", "LLDebugTool/Core/Storage/*.{h,m}", "LLDebugTool/Core/Others/**/*.{h,m}"
 #    ss.resources                = "LLDebugTool/Components/Network/**/*.{xib,storyboard,bundle}", "LLDebugTool/Core/Storage/*.{xib,storyboard,bundle}", "LLDebugTool/Core/Others/**/*.{xib,storyboard,bundle}"
+#   ss.dependency                 "FMDB", "~> 2.0"
+# end
+
+  s.subspec 'Log' do |ss|
+    ss.public_header_files      = "LLDebugTool/Core/Component/Log/LLLog.h"
+    ss.source_files             = "LLDebugTool/DebugTool/*.{h,m}", "LLDebugTool/Core/Component/Log/**/*.{h,m}", "LLDebugTool/Core/Storage/*.{h,m}", "LLDebugTool/Core/Others/**/*.{h,m}"
+    ss.resources                = "LLDebugTool/Components/Log/**/*.{xib,storyboard,bundle}", "LLDebugTool/Core/Storage/*.{xib,storyboard,bundle}", "LLDebugTool/Core/Others/**/*.{xib,storyboard,bundle}"
+    ss.dependency                 "FMDB", "~> 2.0"
+  end
+  
+#  s.subspec 'Crash' do |ss|
+#    ss.public_header_files      = "LLDebugTool/Core/Component/Crash/LLCrash.h"
+#    ss.source_files             = "LLDebugTool/DebugTool/*.{h,m}", "LLDebugTool/Core/Component/Crash/**/*.{h,m}", "LLDebugTool/Core/Storage/*.{h,m}", "LLDebugTool/Core/Others/**/*.{h,m}"
+#    ss.resources                = "LLDebugTool/Components/Crash/**/*.{xib,storyboard,bundle}", "LLDebugTool/Core/Storage/*.{xib,storyboard,bundle}", "LLDebugTool/Core/Others/**/*.{xib,storyboard,bundle}"
 #    ss.dependency                 "FMDB", "~> 2.0"
-#    ss.dependency                 "LLDebugTool/AppInfo"
 #  end
 
   s.subspec 'AppInfo' do |ss|
