+++ LLDebugTool/Core/Component/Network/Function/LLNetworkHelper.m
@@ -28,6 +28,8 @@
 #import "LLConfig.h"
 #import "LLTool.h"
 
+#import "LLRouter+Network.h"
+
 static LLNetworkHelper *_instance = nil;
 
 @interface LLNetworkHelper ()
@@ -61,7 +63,7 @@ static LLNetworkHelper *_instance = nil;
 //    if (@available(iOS 13.0, *)) {
 //        return [self.reachability currentReachabilityStatus];
 //    } else {
-        return [self networkStateFromStatebar];
+        return [LLRouter networkStateFromStatebar];
 //    }
 }
 
@@ -104,113 +106,6 @@ static LLNetworkHelper *_instance = nil;
 #pragma clang diagnostic pop
         }
     }
-}
-
-- (LLNetworkStatus)networkStateFromStatebar {
-    __block LLNetworkStatus returnValue = LLNetworkStatusNotReachable;
-    if (![[NSThread currentThread] isMainThread]) {
-        dispatch_sync(dispatch_get_main_queue(), ^{
-            returnValue = [self networkStateFromStatebar];
-        });
-        return returnValue;
-    }
-    UIView * statusBarModern = [LLTool getUIStatusBarModern];
-    if (@available(iOS 13.0, *)) {
-#ifdef __IPHONE_13_0
-        if (statusBarModern) {
-            // _UIStatusBarDataCellularEntry
-            id currentData = [[statusBarModern valueForKeyPath:@"_statusBar"] valueForKeyPath:@"currentData"];
-            id _wifiEntry = [currentData valueForKeyPath:@"wifiEntry"];
-            id _cellularEntry = [currentData valueForKeyPath:@"cellularEntry"];
-            if (_wifiEntry && [[_wifiEntry valueForKeyPath:@"isEnabled"] boolValue]) {
-                // If wifiEntry is enabled, is WiFi.
-                returnValue = LLNetworkStatusReachableViaWiFi;
-            } else if (_cellularEntry && [[_cellularEntry valueForKeyPath:@"isEnabled"] boolValue]) {
-                NSNumber *type = [_cellularEntry valueForKeyPath:@"type"];
-                if (type != nil) {
-                    switch (type.integerValue) {
-                        case 5:
-                            returnValue = LLNetworkStatusReachableViaWWAN4G;
-                            break;
-                        case 4:
-                            returnValue = LLNetworkStatusReachableViaWWAN3G;
-                            break;
-                            //                        case 1: // Return 1 when 1G.
-                            //                            break;
-                        case 0:
-                            // Return 0 when no sim card.
-                            returnValue = LLNetworkStatusNotReachable;
-                        default:
-                            returnValue = LLNetworkStatusReachableViaWWAN;
-                            break;
-                    }
-                }
-            }
-        }
-#endif
-    } else {
-        if ([statusBarModern isKindOfClass:NSClassFromString(@"UIStatusBar_Modern")]) {
-            // For iPhoneX
-            NSArray *children = [[[statusBarModern valueForKeyPath:@"_statusBar"] valueForKeyPath:@"foregroundView"] subviews];
-            for (UIView *view in children) {
-                for (id child in view.subviews) {
-                    if ([child isKindOfClass:NSClassFromString(@"_UIStatusBarWifiSignalView")]) {
-                        returnValue = LLNetworkStatusReachableViaWiFi;
-                        break;
-                    }
-                    if ([child isKindOfClass:NSClassFromString(@"_UIStatusBarStringView")]) {
-                        NSString *originalText = [child valueForKey:@"_originalText"];
-                        if ([originalText containsString:@"G"]) {
-                            if ([originalText isEqualToString:@"2G"]) {
-                                returnValue = LLNetworkStatusReachableViaWWAN2G;
-                            } else if ([originalText isEqualToString:@"3G"]) {
-                                returnValue = LLNetworkStatusReachableViaWWAN3G;
-                            } else if ([originalText isEqualToString:@"4G"]) {
-                                returnValue = LLNetworkStatusReachableViaWWAN4G;
-                            } else {
-                                returnValue = LLNetworkStatusReachableViaWWAN;
-                            }
-                            break;
-                        }
-                    }
-                }
-                if (returnValue != LLNetworkStatusNotReachable) {
-                    break;
-                }
-            }
-        } else {
-            // For others iPhone
-            NSArray *children = [[statusBarModern valueForKeyPath:@"foregroundView"] subviews];
-            int type = -1;
-            for (id child in children) {
-                if ([child isKindOfClass:[NSClassFromString(@"UIStatusBarDataNetworkItemView") class]]) {
-                    type = [[child valueForKeyPath:@"dataNetworkType"] intValue];
-                }
-            }
-            switch (type) {
-                case 0:
-                    returnValue = LLNetworkStatusNotReachable;
-                    break;
-                case 1:
-                    returnValue = LLNetworkStatusReachableViaWWAN2G;
-                    break;
-                case 2:
-                    returnValue = LLNetworkStatusReachableViaWWAN3G;
-                    break;
-                case 3:
-                    returnValue = LLNetworkStatusReachableViaWWAN4G;
-                    break;
-                case 4:
-                    returnValue = LLNetworkStatusReachableViaWWAN;
-                    break;
-                case 5:
-                    returnValue = LLNetworkStatusReachableViaWiFi;
-                    break;
-            }
-        }
-    }
-    
-    return returnValue;
 }
 
 @end
