+++ LLDebugTool/Core/Others/Router/LLRouter.m
@@ -1,0 +1,107 @@
+//
+//  LLRouter.m
+//
+//  Copyright (c) 2018 LLDebugTool Software Foundation (https://github.com/HDB-Li/LLDebugTool)
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in all
+//  copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+//  SOFTWARE.
+
+#import "LLRouter.h"
+
+//#import "LLLogHelper.h"
+
+@implementation LLRouter
+
+#pragma mark - Public
++ (void)setCrashHelperEnable:(BOOL)isEnable {
+    [self setEnable:isEnable className:[self crashHelper]];
+}
+
++ (void)setLogHelperEnable:(BOOL)isEnable {
+    [self setEnable:isEnable className:[self logHelper]];
+}
+
++ (void)setNetworkHelperEnable:(BOOL)isEnable {
+    [self setEnable:isEnable className:[self networkHelper]];
+}
+
++ (void)setAppInfoHelperEnable:(BOOL)isEnable {
+    [self setEnable:isEnable className:[self appInfoHelper]];
+}
+
++ (void)setScreenshotHelperEnable:(BOOL)isEnable {
+    [self setEnable:isEnable className:[self screenshotHelper]];
+}
+
++ (void)logInFile:(NSString *)file function:(NSString *)function lineNo:(NSInteger)lineNo level:(LLConfigLogLevel)level onEvent:(NSString *)onEvent message:(NSString *)message {
+#warning 需要修改
+//    [[LLLogHelper shared] logInFile:file function:function lineNo:lineNo level:level onEvent:onEvent message:message];
+}
+
+#pragma mark - Primary
++ (void)setEnable:(BOOL)isEnable className:(NSString *)className {
+    SEL selector = @selector(setEnable:);
+    id sharedInstance = [self sharedWithClassName:className];
+    if (!sharedInstance) {
+        return;
+    }
+    NSAssert([sharedInstance respondsToSelector:selector], @"Invoke %@ Failed.", NSStringFromSelector(selector));
+    NSMethodSignature *signature = [sharedInstance methodSignatureForSelector:selector];
+    NSInvocation *invocation = [NSInvocation invocationWithMethodSignature:signature];
+    [invocation setTarget:sharedInstance];
+    [invocation setSelector:selector];
+    [invocation setArgument:&isEnable atIndex:2];
+    [invocation invoke];
+}
+
++ (id _Nullable)sharedWithClassName:(NSString *)className {
+    Class cls = NSClassFromString(className);
+    if (!cls) {
+        return nil;
+    }
+    NSAssert([cls respondsToSelector:@selector(shared)], @"Get shared instance failed.");
+    return [cls performSelector:@selector(shared)];
+}
+
+#pragma mark Getters
+/// LLCrashHelper class name.
++ (NSString *)crashHelper {
+    return @"LLCrashHelper";
+}
+
+/// LLLogHelper class name.
++ (NSString *)logHelper {
+    return @"LLLogHelper";
+}
+
+/// LLNetworkHelper class name.
++ (NSString *)networkHelper {
+    return @"LLNetworkHelper";
+}
+
+/// LLAppInfoHelper class name.
++ (NSString *)appInfoHelper {
+    return @"LLAppInfoHelper";
+}
+
+/// LLScreenshotHelper class name.
++ (NSString *)screenshotHelper {
+    return @"LLScreenshotHelper";
+}
+
+@end
