+++ LLDebugTool/Core/Storage/LLStorageManager.h
@@ -108,9 +108,6 @@ typedef void(^LLStorageManagerArrayBlock)(NSArray <LLStorageModel *>* _Nullable 
 - (void)clearDatabaseWithComplete:(LLStorageManagerBoolBlock _Nullable)complete;
 - (void)clearDatabaseWithComplete:(LLStorageManagerBoolBlock _Nullable)complete synchronous:(BOOL)synchronous;
 
-#pragma mark -
-- (void)updateDatabaseWithVersion:(NSString *)version complete:(LLStorageManagerBoolBlock _Nullable )complete;
-
 @end
 
 NS_ASSUME_NONNULL_END
